implemented a netmask for defining network accessibility scope
this helps to define the reachability of the object.
For all local networks the netmask should be 255.255.0.0 if the scope
should be smaller it should be 255.255.255.0. both netmask will make
sure to access only local networks similar the own network IP. This
means if the object has an ip like 192.168.0.1 a 255.255.0.0 will
accept access from the entire range between 192.168.0.0 to
192.168.255.255.