Reduce time-of-check-to-time-of-use issues in fs access
Various calls can throw exceptions if e.g. trying to delete a file which
already doesn't exist, so this wraps all patterns where condition is
checked -> condition is acted on with exception handlers for the case
where the condition changes during function execution. Introduces
mkdirIfNotExists and unlinkIfExists for this purpose